Fix order total rounding (backend)

+++ backend/webshop/src/main/java/eindproject/webshop/service/CheckOutService.java
@@ -82,7 +82,7 @@ public class CheckOutService {
 
             orderItem.calculateLineTotal(product.getPrice());
             total += orderItem.getLineTotal();
-            roundedTotal = (double) Math.round(total * 100 / 100);
+            roundedTotal = Math.round(total * 100) / 100.;
 
             product.setStock(product.getStock()-cartItem.getQuantity());
             order.addOrderItem(orderItem);
